Graphing Stable Pulse
Previously for the stable pulse, we were dividing its graph by 3. This is because we were using as a seed for Newton, 3 times the normal form. This normalizes things back so that the graph of the normal form is correct, but then the graph of the actual solution is wrong. Deleted this code, and then adjusted the legend so that it denotes three times the normal form